fix: don't fail deserializing the document if CVSS can't be parsed
Currently the deserialization of the document fails if CVSS v3 can't be parsed. This can
happen if the CVSS vector string is malformed or parsing library can't properly handle the value
(like temporal scores for CVSS v3).
Instead of failing, we should store the original vector string and the score, so that the user can
get the original values and handle them as needed. This will also allow us to to have the
original CVSS vector strings in the output for versions currently not supported by the CVSS library
and enable future support for CVSS v4 in CSAF more easily.